Simple case checking for loop passthrough in dataflow graph.
When it can be determined that a loop will execute at least once, the dataflow graph doesn't need to include the option for the loop to have never executed. Currently checking for literal values in the For range and While condition. This is trivial but still potential useful in conjunction with constant propogation and arithmetic reduction.