Added parser error for trying to use a keyword as a variable name.
I had to add a bit to the parser in order to differentiate the error state for keywords from that of other non-identifier tokens. The list of keywords that are differentiated is somewhat arbitrary.

Since there's a large diff for parser.messages: the new stuff is at the very end.