Fixed two bugs in autodiff level optimization
Ad did not propogate to function arguments, so return value was always assumed to be data. Now arguments are assumed to be AD variables if they are of an autodiffable type. This is a conservative approximation - ideally we would search through function calls to determine which arguments actually may be AD variables.

There was a bug introduced by using shadowed variables. It resulted in converting the result type of an indexed expression into an array type.