Reworked how the Block works in the flowgraph to fix multiple scope issues, now proping subexprs of impure exprs
* Subexpressions of impure expressions can now be propagated by LCM
e.g. in the expression normal_rng(sqrt(123)+2, 1), the whole expression cannot move because it's impure, but sqrt(123)+2 can be pulled out for reuse
* Reworked block in the flowgraph
Take expression propagation in this example:

 1  transformed data {
 2    real z = 1;
 3    real y;
 4    {
 5      real x = normal_rng(123, 1);
 6      z = x * 2;
 7      y = normal_rng(z, 1);
 8    }
 9    y = normal_rng(z, 1);
10  }

x * 2 is an expression that we want to propagate forward to lines 7 and 9.
We can propagate safely to 7, but if we go to 9, we'll reference x, which is out of scope outside of the block. So we need to kill the expression x * 2 in the transfer function at the end of the block.
This is actually not an issue for loops and if-else statements, since they generate a 'meet' in the flowgraph, expressions like x * 2 will be killed by the lattice LUB.
This is likely not only an issue for expression propagation. It's probably also an issue for e.g. copy propagation and reaching definitions (I haven't tested this)

My current solution is to change how blocks behave in the flowgraph. Blocks are now successors to their bodies, rather than predecessors. That means that the 'Block' node will be visited after the block body in the MFP. This lets transfer functions (like expr propagation) explicitly kill off variables that were declared in the block.

This change shouldn't mess too much up, since optimizations tend not to rely on the order that they visit Blocks. The only exception seems to be lazy code motion. To support the original prefix-Block flowgraph, I added an optional flag to the flowgraph function that builds the flowgraph as before.