Optimization debugging (#522)
* WIP - optimization experiments

* Fixed two bugs in autodiff level optimization

Ad did not propogate to function arguments, so return value was always assumed to be data. Now arguments are assumed to be AD variables if they are of an autodiffable type. This is a conservative approximation - ideally we would search through function calls to determine which arguments actually may be AD variables.

There was a bug introduced by using shadowed variables. It resulted in converting the result type of an indexed expression into an array type.

* Disallowed dead code elimination from touching any global vars, fixing an issue where it removed a global initialization.

This was actually caused by code the was insterted in codegen and was never represented in the MIR. That breaks the separation of concern in the compiler.

There are also issues with constant propagation not propagating to the parameters block, and DCE not checking the parameters block for liveness.

* dune promote

* Fixed a bug with liveness analysis and else-less if statements (includes debug code)

Variables could be killed by if-else statement if they were assigned in the if body and had no else.

* Fixed an issue where inlines functions would not simulate return statements

When a function is inlined, the strategy here is to wrap the function body
   in a dummy loop, then replace
   returns with breaks. One issue is early return from internal loops - in
   those cases, a break would only break out of the inner loop. The solution is
   a flag variable to indicate whether a 'return' break has been called, and
   then to check if that flag is set after each loop. Then, if a 'return' break
   is called from an inner loop, there's a cascade of breaks all the way out of
   the dummy loop.

* Fixed how function inlining handles variable name replacement

It was conflicting with the dummy variables introduced by inlining.

* Copy propagation no longer propagates assignments to global variables, similar to DCE

* For DCE, globals are always alive.

It's hard to predict when they'll be used, and the division in the blocks introduces additional complications. I think this isn't an unreasonable model, but slightly better would be livening them at the end of each Stan block.

* Changed the way that inlining reassigns variable names.

Variable names were being replaced globally across the function body, so sometimes names would collide. For example, when two for-loops in series both use 'i' as their index, the first loop index would get replaced adding i => sym1__ to the map, then the second would add i=>sym2__ to the map, replacing the first one. Then 'i' would be replaced by sym2__ in both loops, but it hadn't been declared in the first loop.
Instead, variables with the same name should be replaced by the same new variable name, they shouldn't overwrite eachother in the map. These won't collide since they didn't collide in the original function body. However, this means that two different variables with different types may both be added to the map, so the map should only keep track of string->string name conversions, not string->expr variable conversions.
Along the way I added code to support name->name substitutions and reduced code duplication in mir_utils.

* Fixed issue with expression prop - FnReadData has side effects

* Expression prop - Added additional builtin function to side effects

* Added to side effect list, fixed expr prop kills, fixed sym collisions, cleaned up output

* Added to side effect list
_rng functions effectively have side effects - this could more aptly be 'impure expressions'
* Fixed expr prop kills
Expression propagation was not properly killing off expressions when they were updated by assignments or had side effects
* Fixed sym collisions
Since the symbol generator is reset in between initial MIR production and optimization application, symbols could sometimes collide. Now optimizations produce prefixed symbols, one for each phase.
* Cleaned up output
Removed trailing elses, empty blocks produced by loop unrolling and lazy code motion

* Fixed an issue where expression prop didn't kill stale expressions

* Reworked how the Block works in the flowgraph to fix multiple scope issues, now proping subexprs of impure exprs

* Subexpressions of impure expressions can now be propagated by LCM
e.g. in the expression normal_rng(sqrt(123)+2, 1), the whole expression cannot move because it's impure, but sqrt(123)+2 can be pulled out for reuse
* Reworked block in the flowgraph
Take expression propagation in this example:

 1  transformed data {
 2    real z = 1;
 3    real y;
 4    {
 5      real x = normal_rng(123, 1);
 6      z = x * 2;
 7      y = normal_rng(z, 1);
 8    }
 9    y = normal_rng(z, 1);
10  }

x * 2 is an expression that we want to propagate forward to lines 7 and 9.
We can propagate safely to 7, but if we go to 9, we'll reference x, which is out of scope outside of the block. So we need to kill the expression x * 2 in the transfer function at the end of the block.
This is actually not an issue for loops and if-else statements, since they generate a 'meet' in the flowgraph, expressions like x * 2 will be killed by the lattice LUB.
This is likely not only an issue for expression propagation. It's probably also an issue for e.g. copy propagation and reaching definitions (I haven't tested this)

My current solution is to change how blocks behave in the flowgraph. Blocks are now successors to their bodies, rather than predecessors. That means that the 'Block' node will be visited after the block body in the MFP. This lets transfer functions (like expr propagation) explicitly kill off variables that were declared in the block.

This change shouldn't mess too much up, since optimizations tend not to rely on the order that they visit Blocks. The only exception seems to be lazy code motion. To support the original prefix-Block flowgraph, I added an optional flag to the flowgraph function that builds the flowgraph as before.

* Fixed an issue where copy prop didn't kill assigned vars

* Copy prop now kills local decls when exiting block - possibly unnecessary

* dune promote

* Fixed when expr prop didn't kill expressions due to shadowing

In assignments like `x = f(x)`, expr prop would kill replacements involving x and then add x -> f(x). Instead it should add x -> f(x) and then kill, or equivalently see that x is in f(x) and not add the replacement.

All examples are compiling and correct, horray!

* removed most optimization debug code, restored the fma partial evaluation

* Added integration tests for optimizations. Should setup better.

* Separated the notion of duplicatable and deletable expressions, which were 'side effecting' expressions

Expression propagation needs to know if an expression can be duplicated, and DCE and LCM need to know if an expression can be removed (or done fewer times). Treating them both as the same set was a conservative approximation.

* Adding forgotten test files

* reverted optimization CLI interface

* Removed debug code

* Workaround for issue with fdbody=Skip

* Fixed subtle codegen assumption; added debugging flag to disable optimizations which numerically change the program

the preserve_stability flag is in Optimize.ml, Partial_eval.ml, and Monotone_framework.ml.
I will change this interface, but right now it's only important for debugging.
It's important because we use numerical comparisons to check correctness, which are thrown off by small numerical differences.
Those differences occur with partial optimizations to "composed functions", and also it turns out moving around autodiffable variables in seemingly benign ways.